apps/vision-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch
transformers
Pillow
python-multipart
numpy